customized TableViewAdapter inherits from new generic TableViewAdapterBase<Pojo>
